% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winch-package.R
\docType{package}
\name{winch-package}
\alias{winch}
\alias{winch-package}
\title{winch: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://krlmlr.github.io/r-prof/}
  \item \url{https://github.com/krlmlr/r-prof}
  \item Report bugs at \url{https://github.com/krlmlr/r-prof/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
