% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trace_back.R
\name{winch_add_trace_back}
\alias{winch_add_trace_back}
\title{Enrich an rlang traceback with details on native calls}
\usage{
winch_add_trace_back(trace = rlang::trace_back(bottom = parent.frame()))
}
\arguments{
\item{trace}{An rlang traceback as returned by \code{\link[rlang:trace_back]{rlang::trace_back()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function should be reimplemented in rlang,
to avoid a soft dependency between rlang's traceback format and this package.
}
\keyword{internal}
