% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{winch_available}
\alias{winch_available}
\title{Are native tracebacks available?}
\usage{
winch_available()
}
\value{
A scalar logical.
}
\description{
Returns \code{TRUE} if \code{\link[=winch_trace_back]{winch_trace_back()}} is supported on this platform.
}
\examples{
winch_available()
}
