% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.R
\name{winch_stop}
\alias{winch_stop}
\title{Raise an error from native code}
\usage{
winch_stop(message)
}
\arguments{
\item{message}{The error message.}
}
\value{
This function throws an error and does not return.
}
\description{
Primarily intended for testing.
}
\examples{
try(winch_stop("Test"))
}
\seealso{
\code{\link[=winch_call]{winch_call()}}
}
