% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winch-package.R
\docType{package}
\name{winch-package}
\alias{winch}
\alias{winch-package}
\title{winch: Portable Native and Joint Stack Traces}
\description{
Obtain the native stack trace and fuse it with R's stack trace for easier debugging of R packages with native code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-prof.github.io/winch/}
  \item \url{https://github.com/r-prof/winch}
  \item Report bugs at \url{https://github.com/r-prof/winch/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Ian Lance Taylor (Bundled libbacktrace library)
}

Other contributors:
\itemize{
  \item R Consortium [funder]
  \item Free Software Foundation (Bundled libbacktrace library) [copyright holder]
}

}
\keyword{internal}
