% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{winch_call}
\alias{winch_call}
\title{Call an R function from native code}
\usage{
winch_call(fun, env = parent.frame())
}
\arguments{
\item{fun}{A function callable without arguments.}

\item{env}{The environment in which to evaluate the function call.}
}
\value{
The return value of \code{fun()}.
}
\description{
Primarily intended for testing.
}
\examples{
foo <- function() {
  winch_call(bar)
}

bar <- function() {
  writeLines("Hi!")
}

foo()
}
\seealso{
\code{\link[=winch_stop]{winch_stop()}}
}
